+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - add engine config.
 - add date filters.
+- add and change filters on path, info, value
 
 ## [0.2.0] - 2022-04-03